+++ CHANGELOG.md
@@ -85,3 +85,5 @@
 * Define `Some` pattern synonym
 * Define `None` pattern synonym
 * Define `U_1_II` data type
+* Define `U_I_UU_III_U_II_I` data type
+* Define `Reference` type alias